store:move tikv/gcworker to store/gcworker (#22678)
